MGMT-18592: Add security alerts dashboard (#1067)
* MGMT-18592: Add security alerts dashboard

This patch adds a new dashboard that shows a summary of security alerts
generated by Stackrox instances running on managed hubs.

Currently the patch only contains the changes needed in the manager
database (schema and table), the Go code for the new _syncer_ inside the
manager, and the Grafana dashboard. The agent will also need to be
changed to collect the data, but that is not part of this patch.

Related: https://issues.redhat.com/browse/MGMT-18592
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* MGMT-18592: Add Stackrox integration annotation

This patch adds a new
`global-hub.open-cluster-management.io/with-stackrox-integration`
annotation that will be added to the `multiclusterglobalhub` custom
resource to enable the integration with Stackrox, in particular the new
_Security Alert Counts_ dashboard.

Related: https://issues.redhat.com/browse/MGMT-18592
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

---------

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>